todayWorkList 추가 및 데이터구조 세팅
