+++ exercise_app/src/App.js
@@ -1,10 +1,11 @@
 import './App.css';
 import React from 'react';
-import WorkListCalender from './components/workListCalender';
+import WorkListCalender from './components/todayWorkList';
+import TodayWorkList from './components/todayWorkList';
 
 function App() {
   return (
-    <WorkListCalender />
+    <TodayWorkList />
   );
 }
