+++ exercise_app/src/components/todayWorkList.css
@@ -1,0 +1,20 @@
+h1{
+    text-align: center;
+    width: 100%;
+    height: 100%;
+}
+h2{
+    text-align: center;
+    width: 100%;
+}
+
+.AddWork{
+    text-align: right;
+}
+
+li{
+    font-size:20px;
+    font-weight: 1000;
+    margin-left: 3%;
+    margin-top: 2%;
+}
