--- exercise_app/src/components/workListCalendar.css
@@ -1,8 +1,0 @@
-h1{
-    text-align: center;
-    width: 100%;
-}
-
-.AddWork{
-    text-align: right;
-}
